+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * [added] Move to using Pandora's JSON API instead of the XMLRPC one in hopes of
           being more stable and requiring fewer updates
 * [added] After creating a station, immediately being playback of the station
+* [fixed] Always make sure that toolbar items are enabled when a new song plays
 
 # Version 1.0.22 (4/27/12)